requirements.txt
requests~=2.31.0
flask~=3.0.0
packaging~=23.1
scipy~=1.10.1
yacs~=0.1.8
torcheval~=0.0.7
torchmetrics~=1.1.2
scikit-learn~=1.3.1
matplotlib~=3.7.2
tqdm~=4.66.1
opencv-python~=4.8.0.76